Make kppp check for ppp availablity in the kernel.  I'm not sure what should actually be done upon failure, but it prints out something on stderr.  Right now this won't compile on FreeBSD < 3.0, but it will attempt to load the ppp module if ppp support isn't present.
Now, the code is properly ifdef'd so it shouldn't break on Linux, hell I
started to look at doing the same on Linux, but man.  That code is scary,
and might I add 10x as long.  And that's just to check if ppp support was
loaded into the kernel.  Eeek.

svn path=/trunk/kdenetwork/kppp/; revision=39138